bug: fix for crashing when file not exist before deleting audio file

+++ src/modules/process/ScriptReadingEvaluator.py
@@ -249,8 +249,11 @@ class ScriptReadingEvaluator:
             return False
 
         finally:
-            delete_file(filename)
-            delete_file(converted_audio_path)
+            if os.path.exists(filename):
+                delete_file(filename)
+
+            if os.path.exists(converted_audio_path):
+                delete_file(converted_audio_path)
             
     def calculate_remarks(self, pronunciation: float, wpm_category: int, similarity_score: float, pitch_consistency: int, pacing_score: int, fluency: int):
         """calculating remarks"""
